Make MinimalConf fixture for widget tests
