+++ test/widgets/test_crashme.py
@@ -30,20 +30,10 @@ from libqtile.command.interface import CommandException
 from libqtile.widget.crashme import _CrashMe
 
 
-class MinimalConf(libqtile.confreader.Config):
-    auto_fullscreen = False
-    keys = []
-    mouse = []
-    groups = [libqtile.config.Group("a")]
-    layouts = [libqtile.layout.stack.Stack(num_stacks=1)]
-    floating_layout = libqtile.resources.default_config.floating_layout
-    screens = []
-
-
-def test_crashme_init(manager_nospawn):
+def test_crashme_init(manager_nospawn, minimal_conf_noscreen):
     crash = _CrashMe()
 
-    config = MinimalConf
+    config = minimal_conf_noscreen
     config.screens = [
         libqtile.config.Screen(
             top=libqtile.bar.Bar([crash], 10)
